Create multiple Flow configs from a template
