+++ .gitignore
@@ -1,6 +1,6 @@
 .DS_STORE
 node_modules
-.flowconfig
+scripts/flow/*/.flowconfig
 *~
 *.pyc
 .grunt
