+++ scripts/flow/config/flowconfig
@@ -1,25 +1,25 @@
 [ignore]
-
-<PROJECT_ROOT>/fixtures/.*
-<PROJECT_ROOT>/build/.*
-<PROJECT_ROOT>/scripts/bench/.*
+.*/scripts/bench/.*
 
 # These shims are copied into external projects:
-<PROJECT_ROOT>/scripts/rollup/shims/facebook-www/.*
-<PROJECT_ROOT>/scripts/rollup/shims/react-native/.*
+.*/rollup/shims/facebook-www/.*
+.*/rollup/shims/react-native/.*
 
-<PROJECT_ROOT>/.*/node_modules/y18n/.*
-<PROJECT_ROOT>/node_modules/chrome-devtools-frontend/.*
-<PROJECT_ROOT>/node_modules/devtools-timeline-model/.*
-<PROJECT_ROOT>/node_modules/create-react-class/.*
-<PROJECT_ROOT>/.*/__mocks__/.*
-<PROJECT_ROOT>/.*/__tests__/.*
+.*/node_modules/y18n/.*
+.*/node_modules/chrome-devtools-frontend/.*
+.*/node_modules/devtools-timeline-model/.*
+.*/node_modules/create-react-class/.*
+.*/__mocks__/.*
+.*/__tests__/.*
 
 [include]
+../../../node_modules/
+../../../packages/
+../../../scripts/
 
 [libs]
-./node_modules/fbjs/flow/lib/dev.js
-./scripts/flow
+../../../node_modules/fbjs/flow/lib/dev.js
+../environment.js
 
 [lints]
 untyped-type-import=error
